requirements.txt
telethon>=1.36.0
python-telegram-bot>=21.0
MetaTrader5>=5.0.45
python-dotenv>=1.0.0
